start enabling pause, resume; refactor operation modals to make them more manageable
