+++ islands/customNodes.tsx
@@ -138,7 +138,7 @@ export const GroupNode = ({ data }: { data: NodeData }) => {
       onMouseOver={() => setHover()}
       onMouseLeave={() => resetHover()}
     >
-      <div id="dragHandle" class="flex flex-row items-center py-2">
+      <div id="dragHandle" class="flex flex-row items-center pt-2">
         <IconGripVertical class="w-6 h-6 mx-2 text-purple-100 bg-white border border-purple-200" />
         {producer
           ? <ProducerIcon class="w-5 h-5 mr-2" />
@@ -168,7 +168,7 @@ export const GroupNode = ({ data }: { data: NodeData }) => {
 };
 
 export const OperationNode = (
-  { operation, css }: { operation: Operation; css: string },
+  { operation }: { operation: Operation },
 ) => {
   const key = audienceKey(operation.audience);
   const highlight = opModal.value?.audience &&
@@ -181,7 +181,7 @@ export const OperationNode = (
       type="button"
       class={`flex items-center justify-between w-[260px] h-[64px] group bg-white rounded-lg shadow-lg ${
         highlight ? "border-2 border-purple-600" : "border-1 border-purple-200"
-      } pl-1 pr-2 ${css}`}
+      } pl-1 pr-2 mt-2`}
     >
       <div
         class="whitespace-nowrap text-ellipsis overflow-hidden w-full"
