+++ routes/_layout.tsx
@@ -2,12 +2,13 @@ import { LayoutContext } from "$fresh/server.ts";
 import ServiceMapComponent from "../islands/serviceMap.tsx";
 import { NavBar } from "../islands/nav.tsx";
 import { ReactFlowProvider } from "reactflow";
-import OpModal from "../islands/opModal.tsx";
+import Operation from "../islands/drawer/operation.tsx";
 import { serviceSignal } from "../components/serviceMap/serviceSignal.ts";
 import { initAllServices } from "../lib/fetch.ts";
 import { GRPC_TOKEN } from "../lib/configs.ts";
 import { CustomError } from "../components/error/custom.tsx";
 import { Partial } from "$fresh/runtime.ts";
+import { InfoDrawer } from "../islands/drawer/infoDrawer.tsx";
 
 const tokenError = () => (
   <CustomError
@@ -43,7 +44,7 @@ export default async function Layout(req: Request, ctx: LayoutContext) {
     <>
       <NavBar />
       {!req.url.includes("/email") && (
-        <OpModal serviceMap={serviceSignal.value} />
+        <InfoDrawer serviceMap={serviceSignal.value} />
       )}
       <div className="flex flex-col w-screen text-web">
         <Partial name="main-content">
